build: rename onboarding directory to tutorials
