+++ lib/pangea/tutorials/tutorial_constants.dart
