+++ lib/pangea/tutorials/tutorial_overlay_controller.dart
@@ -1,17 +1,15 @@
 import 'dart:async';
 
-import 'package:flutter/material.dart';
-
-import 'package:matrix/matrix_api_lite/utils/logs.dart';
-
 import 'package:fluffychat/pangea/common/utils/firebase_analytics.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_constants.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_enum.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_model.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_overlay_widget.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_sequences.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_state_transition_events.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_constants.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_enum.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_model.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_overlay_widget.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_sequences.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_state_transition_events.dart';
 import 'package:fluffychat/widgets/matrix.dart';
+import 'package:flutter/material.dart';
+import 'package:matrix/matrix_api_lite/utils/logs.dart';
 
 class TutorialOverlayState {
   /// The index next / current tutorial
