+++ test/pangea/tutorial_state_test.dart
@@ -1,10 +1,9 @@
+import 'package:fluffychat/pangea/tutorials/tutorial_enum.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_model.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_overlay_controller.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_state_transition_events.dart';
+import 'package:fluffychat/pangea/tutorials/tutorial_step_model.dart';
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:fluffychat/pangea/onboarding/tutorial_enum.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_model.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_overlay_controller.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_state_transition_events.dart';
-import 'package:fluffychat/pangea/onboarding/tutorial_step_model.dart';
 
 // Sequences used across tests.
 // readingAssistance: 1 step  |  selectModeButtons: 3 steps  |  writingAssistance: 1 step
